+++ challenges/urls.py
@@ -12,6 +12,9 @@ urlpatterns = [
     path('<slug:slug>/submit-challenge-form-1/', views.submit_challenge, name='submit_challenge'),
     path('<slug:slug>/submit-challenge-successful/', views.submit_challenge_successful, name='submit_challenge_successful'),
     # path('add-category/', views.add_category_view.as_view(), name='add_category'),
+    path('<slug:slug>/management-dashboard/', views.PostList.as_view(), name='dashboard'),
+    path('<slug:slug>/completed-challenges/', views.PostListCompleted.as_view(), name='completed_challenges'),
+    path('<slug:slug>/pending-ideas/', views.PendingIdeasList.as_view(), name='pending_ideas'),
     path('<slug:slug>/statistics/', views.Statistics.as_view(), name='stats'),
 
     path('challenge-history/completed-challenges/', views.HistoryListCompleted.as_view(), name='history_completed'),
@@ -22,6 +25,9 @@ urlpatterns = [
     path('<slug:slug>/challenge-history/ideas/delivered/<int:pk>', views.IdeaListDelivered.as_view(), name='delivered_ideas'),
     path('<slug:slug>/challenge-history/', views.History.as_view(), name='challenge_history'),
     path('<slug:slug>/challenge-history/lifecycle/<int:pk>', views.lifecycle, name='lifecycle'),
+    path('<slug:orgslug>/<int:pk>/<slug:slug>/manage-ideas/', views.IdeaManagementDetail.as_view(), name='idea_management_detail'),
+    path('<slug:orgslug>/<int:pk>/<slug:slug>/manage/approve', views.approve_idea, name='approve_idea'),
+    path('<slug:orgslug>/<int:pk>/manage/<slug:slug>/reject', views.reject_idea, name='reject_idea'),
     path('router/', include(router.urls)),
     path('selected-ideas/', views.IdeaViewSet, name='selected_ideas'),
     path('api-auth/', include('rest_framework.urls', namespace='rest_framework')),
