+++ templates/blogs/manager_index.html
@@ -1,4 +1,13 @@
-{% extends "blogs/base.html" %} 
+{% extends 'blogs/index_base.html' %}
+
+{%block nav_bar%} {% include 'components/navbar_management.html' %} {%endblock%}
+
+{%block section_title%}Pending Challenges{%endblock%}
+
+{%block button%}
+<a href="{% url 'post_management_detail' orgslug post.pk post.slug %}" class="nhsuk-button">Manage</a>
+{%endblock button%}
+{% comment %} {% extends "blogs/base.html" %} 
 {% block content %}
 
 <style>
@@ -20,14 +29,15 @@
 <div class="container">
     <br>
     <div class="nhsuk-back-link">
-        <a class="nhsuk-back-link__link" href="{% url 'blogfeed_main' org.slug %}">
+        <a class="nhsuk-back-link__link" href="{% url 'blogfeed_main' orgslug %}">
           <svg class="nhsuk-icon nhsuk-icon__chevron-left" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" aria-hidden="true">
             <path d="M8.5 12c0-.3.1-.5.3-.7l5-5c.4-.4 1-.4 1.4 0s.4 1 0 1.4L10.9 12l4.3 4.3c.4.4.4 1 0 1.4s-1 .4-1.4 0l-5-5c-.2-.2-.3-.4-.3-.7z"></path>
           </svg>
           Go back</a>
-      
+    
       </div>
-    <h1>Completed Challenges</h1>
+    
+    <h1>Pending Challenges</h1>
     <div class="row">
         <!-- Blog Entries Column -->
         <div class="col-md-8 mt-3 left">
@@ -36,15 +46,12 @@
             <div class="card mb-4">
                 <div class="card-body">
                     <h2 class="card-title">{{ post.title }}</h2>
-                    <p class="card-text text-muted h6">Author: {{ post.author|default_if_none:"Anonymous" }} | Date: {{ post.created_on}} </p>
-                    <p class="card-text text-muted h6">Winner: {{ post.winner|default_if_none:"Anonymous" }}</p>
+                    <p class="card-text text-muted h6">{{ post.author }} | {{ post.created_on}} </p>
                     <br>
                     <p class="card-text">{{post.description|slice:":200" }}...</p>
                     <p class="card-text"> Start Date: {{post.startDate }} <br> End Date: {{post.endDate  }}</p>
                     <p class="card-text">                                 Severity: {{post.severity|capfirst}} </p>
-
-
-                    <a href="{% url 'post_management_detail' org.slug post.pk post.slug %}" class="nhsuk-button">Read More</a>
+                    <a href="{% url 'post_management_detail' orgslug post.pk post.slug %}" class="nhsuk-button">Manage</a>
                 </div>
             </div>
             {% endfor %}
@@ -54,21 +61,6 @@
         </div>
         {% block sidebar %} {% include 'components/sidebar.html' %} {% endblock sidebar %}
     </div>
-    {% if is_paginated %}
-    <div class="pagination">
-        <span class="page-links">
-            {% if page_obj.has_previous %}
-            <a href="/challenges/{{ org.slug }}/completed-challenges/?page={{ page_obj.previous_page_number }}" class="nhsuk-button">Previous Page</a>
-            {% endif %}
-            {% if page_obj.has_next %}
-                <a href="/challenges/{{ org.slug }}/completed-challenges/?page={{ page_obj.next_page_number }}" class="nhsuk-button">Next Page</a>
-            {% endif %}
-            <span class="page-current">
-                <h2 class="card-title">Page {{ page_obj.number }} of {{ page_obj.paginator.num_pages }}</h2>                    
-            </span>
-        </span>
-    </div>
-{% endif %}
 </div>
 {% include 'components/footer.html' %}
-{%endblock%}
+{%endblock%} {% endcomment %}
