+++ templates/blogs/post_management_detail.html
@@ -41,7 +41,7 @@
       <div class="card-body">
 
         <div class="nhsuk-back-link">
-  <a class="nhsuk-back-link__link" href="{% url 'dashboard' org.slug %}">
+  <a class="nhsuk-back-link__link" href="{% url 'dashboard' orgslug %}">
     <svg class="nhsuk-icon nhsuk-icon__chevron-left" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" aria-hidden="true">
       <path d="M8.5 12c0-.3.1-.5.3-.7l5-5c.4-.4 1-.4 1.4 0s.4 1 0 1.4L10.9 12l4.3 4.3c.4.4.4 1 0 1.4s-1 .4-1.4 0l-5-5c-.2-.2-.3-.4-.3-.7z"></path>
     </svg>
@@ -76,9 +76,9 @@
 
 {% if is_active %}
 
-<button formaction="{% url 'approval_view' org.slug post.pk post.slug %}" type="submit" name="post_id" value="{{ post.id }}" class="nhsuk-button">Review</button>
+<button formaction="{% url 'approval_view' orgslug post.pk post.slug %}" type="submit" name="post_id" value="{{ post.id }}" class="nhsuk-button">Review</button>
 
-<button formaction="{% url 'reject_challenge' org.slug post.pk post.slug %}" id="submit-button" type="submit" name="post_id" value="{{ post.id }}" class="nhsuk-button nhsuk-button--secondary" onclick="return confirm('Are you sure you want to reject this item?');">Reject</button>
+<button formaction="{% url 'reject_challenge' orgslug post.pk post.slug %}" id="submit-button" type="submit" name="post_id" value="{{ post.id }}" class="nhsuk-button nhsuk-button--secondary" onclick="return confirm('Are you sure you want to reject this item?');">Reject</button>
 
 {% else %}
 {% endif %}
