+++ templates/components/navbar_management.html
@@ -74,7 +74,7 @@
   <header class="nhsuk-header" role="banner">
   <div class="nhsuk-width-container nhsuk-header__container">
     <div class="nhsuk-header__logo">
-  <a class="nhsuk-header__link" href="{% url 'blogfeed_main' org.slug %}">
+  <a class="nhsuk-header__link" href="{% url 'blogfeed_main' orgslug %}">
     <image style="height: 50px; width: 110px; border-radius: 10px; " src="{% static 'frontend/imgs/ixn2.png' %}" xlink:href=""></image>
 
     
@@ -92,7 +92,7 @@
 
         <div class="nhsuk-header__search">
             <div class="nhsuk-header__search-wrap beta-header__search-wrap js-show" id="wrap-search">
-              <form class="nhsuk-header__search-form beta-header__search-form" id="search" action="{% url 'search_blog' org.slug %}" method="post" role="search">
+              <form class="nhsuk-header__search-form beta-header__search-form" id="search" action="{% url 'search_blog' orgslug %}" method="post" role="search">
                 {% csrf_token %}
                 <label class="nhsuk-u-visually-hidden" for="search-field">Search the NHS website</label>
                 <div class="autocomplete-container" id="autocomplete-container"></div>
@@ -123,7 +123,7 @@
   </p>
   <ul class="nhsuk-header__navigation-list">
     <li class="nhsuk-header__navigation-item nhsuk-header__navigation-item--for-mobile">
-      <a class="nhsuk-header__navigation-link" href="{% url 'blogfeed_main' org.slug %}" >
+      <a class="nhsuk-header__navigation-link" href="{% url 'blogfeed_main' orgslug %}" >
         Home
         <svg class="nhsuk-icon nhsuk-icon__chevron-right" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" aria-hidden="true">
           <path d="M15.5 12a1 1 0 0 1-.29.71l-5 5a1 1 0 0 1-1.42-1.42l4.3-4.29-4.3-4.29a1 1 0 0 1 1.42-1.42l5 5a1 1 0 0 1 .29.71z"></path>
