--- templates/management/manager_index.html
@@ -1,66 +1,0 @@
-{% extends 'blogs/index_base.html' %}
-
-{%block nav_bar%} {% include 'components/navbar_management.html' %} {%endblock%}
-
-{%block section_title%}Pending Challenges{%endblock%}
-
-{%block button%}
-<a href="{% url 'post_management_detail' org.slug post.pk post.slug %}" class="nhsuk-button">Manage</a>
-{%endblock button%}
-{% comment %} {% extends "blogs/base.html" %} 
-{% block content %}
-
-<style>
-    body {
-        font-family: "Roboto", sans-serif;
-        font-size: 18px;
-        background-color: #fdfdfd;
-    }
-    
-    .head_text {
-        color: white;
-    }
-    
-    .card {
-        box-shadow: 0 16px 48px #E3E7EB;
-    }
-</style>
-
-<div class="container">
-    <br>
-    <div class="nhsuk-back-link">
-        <a class="nhsuk-back-link__link" href="{% url 'blogfeed_main' orgslug %}">
-          <svg class="nhsuk-icon nhsuk-icon__chevron-left" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" aria-hidden="true">
-            <path d="M8.5 12c0-.3.1-.5.3-.7l5-5c.4-.4 1-.4 1.4 0s.4 1 0 1.4L10.9 12l4.3 4.3c.4.4.4 1 0 1.4s-1 .4-1.4 0l-5-5c-.2-.2-.3-.4-.3-.7z"></path>
-          </svg>
-          Go back</a>
-    
-      </div>
-    
-    <h1>Pending Challenges</h1>
-    <div class="row">
-        <!-- Blog Entries Column -->
-        <div class="col-md-8 mt-3 left">
-            {% if post_list %}
-            {% for post in post_list %}
-            <div class="card mb-4">
-                <div class="card-body">
-                    <h2 class="card-title">{{ post.title }}</h2>
-                    <p class="card-text text-muted h6">{{ post.author }} | {{ post.created_on}} </p>
-                    <br>
-                    <p class="card-text">{{post.description|slice:":200" }}...</p>
-                    <p class="card-text"> Start Date: {{post.startDate }} <br> End Date: {{post.endDate  }}</p>
-                    <p class="card-text">                                 Severity: {{post.severity|capfirst}} </p>
-                    <a href="{% url 'post_management_detail' orgslug post.pk post.slug %}" class="nhsuk-button">Manage</a>
-                </div>
-            </div>
-            {% endfor %}
-            {% else %}
-            <h3>There are currently no pending polls - you will be notified once a user makes a request</h3>
-            {% endif %}
-        </div>
-        {% block sidebar %} {% include 'components/sidebar.html' %} {% endblock sidebar %}
-    </div>
-</div>
-{% include 'components/footer.html' %}
-{%endblock%} {% endcomment %}
